refactor(cli): restore family split size budget
